+++ pytket/extensions/aqt/multi_zone_architecture/graph_algs/__init__.py
@@ -11,25 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import json
-from dataclasses import dataclass
-from typing import Optional, Sequence
-from qiskit_aqt_provider import api_models
-
-from pytket import Circuit
-from pytket.backends import ResultHandle
-
-
-@dataclass
-class PytketAqtJobCircuitData:
-    circuit: Circuit
-    n_shots: int
-    postprocess_json: str = json.dumps(None)
-    aqt_circuit: Optional[api_models.Circuit] = None
-    measures: Optional[str] = None
-    handle: Optional[ResultHandle] = None
-
-
-@dataclass
-class PytketAqtJob:
-    circuits_data: Sequence[PytketAqtJobCircuitData]